refactor(tcp)!: name the extension seam Policy rather than LossRecovery
The interface began as loss recovery and has outgrown the name. Two of its three
implementations are not recovering from loss: the timestamp policy negotiates an
option and samples round-trip times from it, and the window scale offer rides the
same option hook. Both need exactly the hooks loss recovery needs, at exactly the
same points, and the name was the only thing suggesting they did not belong.

Rename the interface, its config field and the handler's accessor accordingly, and
say in the doc why the seam is broader than one of its clients: the core owns the
state machine, the sequence space and the buffers, while a policy adds timing,
windows and options, none of which the state machine needs to be correct. A
connection with no policy still handshakes, carries data and closes; it simply
never retransmits.

Mechanical apart from the interface's own documentation. No behaviour change.

BREAKING CHANGE: tcp.LossRecovery is now tcp.Policy, ConnConfig.LossRecovery is
now ConnConfig.Policy, and Handler.SetLossRecovery is now Handler.SetPolicy.
