+++ tcp/rto/rto.go
@@ -35,12 +35,12 @@ const (
 )
 
 // Timer implements the RFC 6298 round-trip-time estimator and the single
-// retransmission timer as a [tcp.LossRecovery]. Construct it with new(Timer) and hand
-// it to [tcp.ConnConfig.LossRecovery]; the connection calls [Timer.Reset] on open, so
+// retransmission timer as a [tcp.Policy]. Construct it with new(Timer) and hand
+// it to [tcp.ConnConfig.Policy]; the connection calls [Timer.Reset] on open, so
 // the zero value is ready to use.
 //
 // Timer is a pure, reactive state machine: it observes the segments a connection
-// sends and receives (via the tcp.LossRecovery hooks) and the monotonic time handed
+// sends and receives (via the tcp.Policy hooks) and the monotonic time handed
 // in at each hook, and from those alone derives RTT estimates and retransmission
 // decisions. It holds no clock and allocates nothing, which keeps it
 // deterministic for unit testing (see issue #140).
@@ -81,16 +81,16 @@ type Timer struct {
 	expirations uint32
 }
 
-var _ tcp.LossRecovery = (*Timer)(nil)
+var _ tcp.Policy = (*Timer)(nil)
 
 // Reset returns the estimator to its pre-connection state with the initial RTO.
-// It implements [tcp.LossRecovery] and is called when the connection opens or aborts
+// It implements [tcp.Policy] and is called when the connection opens or aborts
 // so the estimator can be reused across connection reuse.
 func (r *Timer) Reset() { *r = Timer{rto: rtoInitial} }
 
 // SmoothedRTT returns the current smoothed round-trip time (SRTT), or zero
 // before the first RTT measurement. It is concrete-type introspection and is
-// intentionally not part of [tcp.LossRecovery].
+// intentionally not part of [tcp.Policy].
 func (r *Timer) SmoothedRTT() time.Duration { return r.srtt }
 
 // CurrentRTO returns the timeout currently in effect, clamped to [rtoMin, rtoMax].
@@ -108,7 +108,7 @@ func (r *Timer) CurrentRTO() time.Duration {
 func (r *Timer) Running() bool { return r.running }
 
 // NextDeadline returns the monotonic-nanosecond instant at which the timer
-// expires, or 0 when it is not armed. It implements [tcp.LossRecovery].
+// expires, or 0 when it is not armed. It implements [tcp.Policy].
 func (r *Timer) NextDeadline() int64 {
 	if !r.running {
 		return 0
@@ -118,13 +118,13 @@ func (r *Timer) NextDeadline() int64 {
 
 // PreRx keeps every segment: the estimator never drops traffic and records
 // nothing before the connection has decided whether the segment counts. It
-// implements [tcp.LossRecovery].
+// implements [tcp.Policy].
 func (r *Timer) PreRx(rx tcp.RxMeta) tcp.RxDirective {
 	return tcp.RxDirective{Keep: true}
 }
 
 // PostRx samples the RTT and manages the retransmission timer from a segment the
-// connection accepted (RFC 6298 §5.2/§5.3). It implements [tcp.LossRecovery].
+// connection accepted (RFC 6298 §5.2/§5.3). It implements [tcp.Policy].
 //
 // A refused segment is ignored. Acting on one would let an acknowledgement the
 // state machine rejected, for data never sent, collapse the backoff and take a
@@ -156,14 +156,14 @@ func (r *Timer) PostRx(event tcp.RxEvent) {
 }
 
 // WriteOptions adds no TCP options: retransmission timing needs none of its
-// own. It implements [tcp.LossRecovery].
+// own. It implements [tcp.Policy].
 func (r *Timer) WriteOptions(plan tcp.TxPlan, opts []byte) uint8 { return 0 }
 
 // PreTx reports whether the retransmission timer has expired and, if so, applies
 // the RFC 6298 §5.4–§5.6 timeout response — discard the outstanding RTT sample
 // (Karn), back the RTO off exponentially and restart the timer — returning a
 // directive that asks the connection to retransmit from snd.UNA (go-back-N). It
-// implements [tcp.LossRecovery].
+// implements [tcp.Policy].
 func (r *Timer) PreTx(intent tcp.TxIntent) tcp.TxDirective {
 	now := intent.Now
 	if !r.running || now < r.deadline || r.sndUNA == r.sndNXT {
@@ -184,7 +184,7 @@ func (r *Timer) PreTx(intent tcp.TxIntent) tcp.TxDirective {
 // begins timing newly transmitted data (RFC 6298 §3) and arms the timer (§5.1).
 // Segments that do not extend the send sequence are retransmissions and are
 // never RTT-sampled (Karn's algorithm). Control-only segments (no data) are
-// ignored. It implements [tcp.LossRecovery].
+// ignored. It implements [tcp.Policy].
 func (r *Timer) PostTx(outgoing tcp.Segment, now int64) {
 	if outgoing.DATALEN == 0 {
 		return // only data segments are timed / arm the RTO.
